% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDomainInfo.R
\name{get_domain_info}
\alias{get_domain_info}
\title{Retrieves all domains hosted on a single IP address}
\usage{
get_domain_info(ip = NULL, page = NULL, limit = NULL)
}
\arguments{
\item{ip}{(required) The IP you want to query the domains for}

\item{page}{The page for iterating through all domains}

\item{limit}{The amount of domains you want to see on each page}
}
\value{
Retrieves all domains hosted on a single IP address
}
\description{
Retrieves all domains hosted on a single IP address
}
